requirements.txt
asgiref==3.5.2
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.0
Django==4.1
exceptiongroup==1.2.2
h11==0.14.0
idna==3.8
outcome==1.3.0.post0
pip==21.2.3
pycparser==2.22
PySocks==1.7.1
python-decouple==3.8
pytz==2022.2.1
selenium==4.24.0
setuptools==57.4.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2022.2
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
djangorestframework==3.13.1
psycopg2-binary==2.9.3